pyproject.toml
[project]
name = "gleasonxai"
dynamic = ["version"]          # tell PEP-621 that setuptools will supply these
description = "Code for the the paper 'Pathologist-like explainable AI for interpretable Gleason grading in prostate cancer'."
readme = "README.md"


authors = [
    { name = "Hendrik Mehrtens", email = "hendrikalexander.mehrtens@dkfz-heidelberg.de" },
    { name = "Gesa Mittmann", email = "gesa.mittmann@dkfz-heidelberg.de"}
]

requires-python = ">=3.10,<3.11"
dependencies = [
    "albumentations===1.3.1",
    "hydra-core===1.3.2",
    "hydra-optuna-sweeper===1.2.0",
    "ipykernel>=6.29.5",
    "ipywidgets===8.1.2",
    "matplotlib===3.8.0",
    "monai===1.3.0",
    "omegaconf>=2.3.0",
    "opencv-python===4.8.1.78",
    "pandas===2.1.1",
    "pillow===9.5.0",
    "pytorch-lightning===2.2.0.post0",
    "pytorchcv>=0.0.73",
    "requests>=2.32.3",
    "scikit-image===0.22.0",
    "scikit-learn===1.3.2",
    "scipy===1.11.3",
    "seaborn===0.13.2",
    "segmentation-models-pytorch===0.3.3",
    "statsmodels===0.14.2",
    "tensorboard===2.16.2",
    "timm===0.9.2",
    "torch===2.1.1",
    "torchaudio===2.1.1",
    "torchmetrics===1.3.0",
    "torchvision===0.16.1",
    "tqdm>=4.67.1",
    "wandb===0.17.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "gleasonxai.__version__"}   # Get the version dynamically


[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
